# MQE: Add support for more aggregation functions