--- a/pkg/streamingpromql/aggregations/common.go
+++ b/pkg/streamingpromql/aggregations/common.go
@@ -21,6 +21,7 @@ type AggregationGroup interface {
 type AggregationGroupFactory func() AggregationGroup
 
 var AggregationGroupFactories = map[parser.ItemType]AggregationGroupFactory{
-	parser.MAX: func() AggregationGroup { return &MaxAggregationGroup{} },
+	parser.MAX: func() AggregationGroup { return &MinMaxAggregationGroup{max: true} },
+	parser.MIN: func() AggregationGroup { return &MinMaxAggregationGroup{max: false} },
 	parser.SUM: func() AggregationGroup { return &SumAggregationGroup{} },
 }