--- a/pkg/streamingpromql/aggregations/common.go
+++ b/pkg/streamingpromql/aggregations/common.go
@@ -22,7 +22,6 @@ type AggregationGroup interface {
 type AggregationGroupFactory func() AggregationGroup
 
 var AggregationGroupFactories = map[parser.ItemType]AggregationGroupFactory{
-	parser.AVG: func() AggregationGroup { return &AvgAggregationGroup{} },
 	parser.MAX: func() AggregationGroup { return &MinMaxAggregationGroup{max: true} },
 	parser.MIN: func() AggregationGroup { return &MinMaxAggregationGroup{max: false} },
 	parser.SUM: func() AggregationGroup { return &SumAggregationGroup{} },