--- a/pkg/streamingpromql/aggregations/common.go
+++ b/pkg/streamingpromql/aggregations/common.go
@@ -22,8 +22,8 @@ type AggregationGroup interface {
 type AggregationGroupFactory func() AggregationGroup
 
 var AggregationGroupFactories = map[parser.ItemType]AggregationGroupFactory{
-	parser.MAX: func() AggregationGroup { return &MinMaxAggregationGroup{max: true} },
-	parser.MIN: func() AggregationGroup { return &MinMaxAggregationGroup{max: false} },
+	parser.MAX: func() AggregationGroup { return NewMinMaxAggregationGroup(true) },
+	parser.MIN: func() AggregationGroup { return NewMinMaxAggregationGroup(false) },
 	parser.SUM: func() AggregationGroup { return &SumAggregationGroup{} },
 }