--- a/pkg/streamingpromql/aggregations/min_max.go
+++ b/pkg/streamingpromql/aggregations/min_max.go
@@ -47,18 +47,18 @@ func (g *MinMaxAggregationGroup) minAccumulatePoint(idx int64, t int64, f float6
 	}
 }
 
-func (g *MinMaxAggregationGroup) AccumulateSeries(data types.InstantVectorSeriesData, steps int, start int64, interval int64, memoryConsumptionTracker *limiting.MemoryConsumptionTracker, _ functions.EmitAnnotationFunc) (bool, error) {
+func (g *MinMaxAggregationGroup) AccumulateSeries(data types.InstantVectorSeriesData, steps int, start int64, interval int64, memoryConsumptionTracker *limiting.MemoryConsumptionTracker, _ functions.EmitAnnotationFunc) error {
 	if len(data.Floats) > 0 && g.floatValues == nil {
 		var err error
 		// First series with float values for this group, populate it.
 		g.floatValues, err = types.Float64SlicePool.Get(steps, memoryConsumptionTracker)
 		if err != nil {
-			return false, err
+			return err
 		}
 
 		g.floatPresent, err = types.BoolSlicePool.Get(steps, memoryConsumptionTracker)
 		if err != nil {
-			return false, err
+			return err
 		}
 		g.floatValues = g.floatValues[:steps]
 		g.floatPresent = g.floatPresent[:steps]
@@ -80,10 +80,10 @@ func (g *MinMaxAggregationGroup) AccumulateSeries(data types.InstantVectorSeries
 	}
 
 	types.PutInstantVectorSeriesData(data, memoryConsumptionTracker)
-	return false, nil
+	return nil
 }
 
-func (g *MinMaxAggregationGroup) ComputeOutputSeries(start int64, interval int64, memoryConsumptionTracker *limiting.MemoryConsumptionTracker) (types.InstantVectorSeriesData, error) {
+func (g *MinMaxAggregationGroup) ComputeOutputSeries(start int64, interval int64, memoryConsumptionTracker *limiting.MemoryConsumptionTracker) (types.InstantVectorSeriesData, bool, error) {
 	floatPointCount := 0
 	for _, p := range g.floatPresent {
 		if p {
@@ -95,7 +95,7 @@ func (g *MinMaxAggregationGroup) ComputeOutputSeries(start int64, interval int64
 	if floatPointCount > 0 {
 		floatPoints, err = types.FPointSlicePool.Get(floatPointCount, memoryConsumptionTracker)
 		if err != nil {
-			return types.InstantVectorSeriesData{}, err
+			return types.InstantVectorSeriesData{}, false, err
 		}
 
 		for i, havePoint := range g.floatPresent {
@@ -110,5 +110,5 @@ func (g *MinMaxAggregationGroup) ComputeOutputSeries(start int64, interval int64
 	types.Float64SlicePool.Put(g.floatValues, memoryConsumptionTracker)
 	types.BoolSlicePool.Put(g.floatPresent, memoryConsumptionTracker)
 
-	return types.InstantVectorSeriesData{Floats: floatPoints}, nil
+	return types.InstantVectorSeriesData{Floats: floatPoints}, false, nil
 }