… CHANGELOG.md
… cmd/mimir/config-descriptor.json
… docs/sources/mimir/configure/configuration-parameters/index.md
… pkg/streamingpromql/aggregations/sum.go
… pkg/streamingpromql/config.go
… pkg/streamingpromql/engine_test.go
… pkg/streamingpromql/operators/aggregation.go
… pkg/streamingpromql/operators/aggregation_test.go
… pkg/streamingpromql/query.go
… pkg/streamingpromql/testdata/ours/aggregators.test
… pkg/streamingpromql/testdata/upstream/aggregators.test