#### What this PR does

- Adds feature flag for disabling aggregations
- Refactors and breaks out aggregation to support different functions
- Implements `max` aggregation function

#### Which issue(s) this PR fixes or relates to

Related to #2280

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be `[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`.
- [ ] [`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md) updated with experimental features.