@jhesketh 2024-08-20
Benchmark vs main:

<details>

```

goos: linux
goarch: amd64
pkg: github.com/grafana/mimir/pkg/streamingpromql/benchmarks
cpu: Intel(R) Xeon(R) CPU E5-2686 v4 @ 2.30GHz
                                                                                                                      │   main.txt   │              agg.txt               │
                                                                                                                      │    sec/op    │    sec/op     vs base              │
Query/sum_over_time(a_1[1m]),_instant_query/Mimir-8                                                                      610.6µ ± 4%   637.0µ ±  2%  +4.32% (p=0.009 n=6)
Query/sum_over_time(a_1[1m]),_instant_query/Prometheus-8                                                                 657.9µ ± 1%   671.1µ ±  6%       ~ (p=0.132 n=6)
Query/sum_over_time(a_1[1m]),_range_query_with_100_steps/Mimir-8                                                         681.7µ ± 6%   687.7µ ±  1%       ~ (p=0.589 n=6)
Query/sum_over_time(a_1[1m]),_range_query_with_100_steps/Prometheus-8                                                    721.2µ ± 1%   740.8µ ±  2%  +2.72% (p=0.002 n=6)
Query/sum_over_time(a_1[1m]),_range_query_with_1000_steps/Mimir-8                                                        892.0µ ± 6%   912.9µ ±  4%       ~ (p=0.240 n=6)
Query/sum_over_time(a_1[1m]),_range_query_with_1000_steps/Prometheus-8                                                   988.2µ ± 1%   994.6µ ±  2%       ~ (p=0.093 n=6)
Query/sum_over_time(a_100[1m]),_instant_query/Mimir-8                                                                    1.272m ± 2%   1.324m ±  1%  +4.07% (p=0.002 n=6)
Query/sum_over_time(a_100[1m]),_instant_query/Prometheus-8                                                               1.437m ± 1%   1.467m ±  4%       ~ (p=0.065 n=6)
Query/sum_over_time(a_100[1m]),_range_query_with_100_steps/Mimir-8                                                       4.208m ± 1%   4.293m ±  1%  +2.03% (p=0.002 n=6)
Query/sum_over_time(a_100[1m]),_range_query_with_100_steps/Prometheus-8                                                  4.937m ± 1%   4.992m ±  3%  +1.11% (p=0.004 n=6)
Query/sum_over_time(a_100[1m]),_range_query_with_1000_steps/Mimir-8                                                      21.05m ± 1%   21.34m ±  1%  +1.34% (p=0.002 n=6)
Query/sum_over_time(a_100[1m]),_range_query_with_1000_steps/Prometheus-8                                                 29.44m ± 2%   29.77m ±  1%       ~ (p=0.065 n=6)
Query/sum_over_time(a_2000[1m]),_instant_query/Mimir-8                                                                   10.28m ± 1%   10.50m ±  1%  +2.10% (p=0.002 n=6)
Query/sum_over_time(a_2000[1m]),_instant_query/Prometheus-8                                                              12.23m ± 2%   12.66m ±  2%  +3.49% (p=0.002 n=6)
Query/sum_over_time(a_2000[1m]),_range_query_with_100_steps/Mimir-8                                                      69.28m ± 1%   69.45m ±  1%       ~ (p=0.394 n=6)
Query/sum_over_time(a_2000[1m]),_range_query_with_100_steps/Prometheus-8                                                 82.20m ± 2%   82.53m ±  1%       ~ (p=0.485 n=6)
Query/sum_over_time(a_2000[1m]),_range_query_with_1000_steps/Mimir-8                                                     398.0m ± 1%   398.9m ±  0%       ~ (p=0.818 n=6)
Query/sum_over_time(a_2000[1m]),_range_query_with_1000_steps/Prometheus-8                                                561.4m ± 1%   562.8m ±  0%       ~ (p=0.394 n=6)
Query/sum_over_time(nh_1[1m]),_instant_query/Mimir-8                                                                     653.6µ ± 2%   671.1µ ±  3%  +2.68% (p=0.009 n=6)
Query/sum_over_time(nh_1[1m]),_instant_query/Prometheus-8                                                                685.5µ ± 1%   700.4µ ±  6%       ~ (p=0.394 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_100_steps/Mimir-8                                                        1.058m ± 1%   1.055m ±  2%       ~ (p=0.937 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_100_steps/Prometheus-8                                                   1.021m ± 5%   1.041m ±  3%       ~ (p=0.065 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_1000_steps/Mimir-8                                                       2.395m ± 2%   2.401m ±  1%       ~ (p=0.240 n=6)
Query/sum_over_time(nh_1[1m]),_range_query_with_1000_steps/Prometheus-8                                                  2.716m ± 2%   2.755m ±  2%       ~ (p=0.093 n=6)
Query/sum_over_time(nh_100[1m]),_instant_query/Mimir-8                                                                   1.739m ± 1%   1.762m ±  1%       ~ (p=0.132 n=6)
Query/sum_over_time(nh_100[1m]),_instant_query/Prometheus-8                                                              2.045m ± 2%   2.078m ±  1%  +1.62% (p=0.015 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_100_steps/Mimir-8                                                      35.66m ± 0%   36.30m ±  1%  +1.77% (p=0.002 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_100_steps/Prometheus-8                                                 29.37m ± 1%   29.73m ±  1%  +1.21% (p=0.002 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_1000_steps/Mimir-8                                                     169.2m ± 1%   170.3m ±  0%       ~ (p=0.093 n=6)
Query/sum_over_time(nh_100[1m]),_range_query_with_1000_steps/Prometheus-8                                                198.2m ± 1%   198.8m ±  1%       ~ (p=0.394 n=6)
Query/sum_over_time(nh_2000[1m]),_instant_query/Mimir-8                                                                  17.71m ± 4%   18.09m ±  7%       ~ (p=0.132 n=6)
Query/sum_over_time(nh_2000[1m]),_instant_query/Prometheus-8                                                             22.35m ± 1%   23.05m ±  1%  +3.12% (p=0.002 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_100_steps/Mimir-8                                                     695.4m ± 1%   697.6m ±  1%       ~ (p=0.589 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_100_steps/Prometheus-8                                                579.8m ± 1%   578.5m ±  2%       ~ (p=0.937 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_1000_steps/Mimir-8                                                     3.580 ± 0%    3.597 ±  1%       ~ (p=0.093 n=6)
Query/sum_over_time(nh_2000[1m]),_range_query_with_1000_steps/Prometheus-8                                                4.191 ± 4%    4.183 ±  0%       ~ (p=0.699 n=6)
Query/sum(a_1),_instant_query/Mimir-8                                                                                    673.5µ ± 3%   670.3µ ±  1%       ~ (p=0.699 n=6)
Query/sum(a_1),_instant_query/Prometheus-8                                                                               667.6µ ± 2%   686.8µ ±  3%  +2.87% (p=0.015 n=6)
Query/sum(a_1),_range_query_with_100_steps/Mimir-8                                                                       688.9µ ± 2%   699.6µ ±  3%  +1.55% (p=0.015 n=6)
Query/sum(a_1),_range_query_with_100_steps/Prometheus-8                                                                  722.7µ ± 3%   714.8µ ±  1%       ~ (p=0.240 n=6)
Query/sum(a_1),_range_query_with_1000_steps/Mimir-8                                                                      848.9µ ± 1%   846.4µ ±  2%       ~ (p=1.000 n=6)
Query/sum(a_1),_range_query_with_1000_steps/Prometheus-8                                                                 933.1µ ± 3%   939.8µ ± 10%       ~ (p=0.394 n=6)
Query/sum(a_100),_instant_query/Mimir-8                                                                                  2.186m ± 1%   2.231m ±  1%  +2.07% (p=0.009 n=6)
Query/sum(a_100),_instant_query/Prometheus-8                                                                             2.250m ± 1%   2.255m ±  2%       ~ (p=0.394 n=6)
Query/sum(a_100),_range_query_with_100_steps/Mimir-8                                                                     3.512m ± 3%   3.599m ±  1%  +2.48% (p=0.026 n=6)
Query/sum(a_100),_range_query_with_100_steps/Prometheus-8                                                                3.546m ± 1%   3.563m ±  1%       ~ (p=0.065 n=6)
Query/sum(a_100),_range_query_with_1000_steps/Mimir-8                                                                    15.83m ± 2%   15.83m ±  1%       ~ (p=0.818 n=6)
Query/sum(a_100),_range_query_with_1000_steps/Prometheus-8                                                               16.12m ± 1%   16.11m ±  5%       ~ (p=0.937 n=6)
Query/sum(a_2000),_instant_query/Mimir-8                                                                                 25.98m ± 1%   25.81m ±  2%       ~ (p=0.699 n=6)
Query/sum(a_2000),_instant_query/Prometheus-8                                                                            26.60m ± 5%   26.81m ±  3%       ~ (p=1.000 n=6)
Query/sum(a_2000),_range_query_with_100_steps/Mimir-8                                                                    51.70m ± 1%   52.51m ±  1%  +1.57% (p=0.002 n=6)
Query/sum(a_2000),_range_query_with_100_steps/Prometheus-8                                                               55.26m ± 2%   54.49m ±  3%       ~ (p=0.180 n=6)
Query/sum(a_2000),_range_query_with_1000_steps/Mimir-8                                                                   277.0m ± 1%   278.0m ±  1%       ~ (p=0.310 n=6)
Query/sum(a_2000),_range_query_with_1000_steps/Prometheus-8                                                              336.8m ± 2%   335.7m ±  1%       ~ (p=0.699 n=6)
Query/sum_without_(l)(h_1),_instant_query/Mimir-8                                                                        781.8µ ± 1%   802.9µ ±  3%  +2.69% (p=0.015 n=6)
Query/sum_without_(l)(h_1),_instant_query/Prometheus-8                                                                   809.8µ ± 2%   820.6µ ±  3%       ~ (p=0.310 n=6)
Query/sum_without_(l)(h_1),_range_query_with_100_steps/Mimir-8                                                           898.6µ ± 3%   890.2µ ±  1%       ~ (p=0.310 n=6)
Query/sum_without_(l)(h_1),_range_query_with_100_steps/Prometheus-8                                                      918.2µ ± 1%   922.0µ ±  2%       ~ (p=0.394 n=6)
Query/sum_without_(l)(h_1),_range_query_with_1000_steps/Mimir-8                                                          1.706m ± 2%   1.718m ±  2%       ~ (p=0.180 n=6)
Query/sum_without_(l)(h_1),_range_query_with_1000_steps/Prometheus-8                                                     1.775m ± 1%   1.803m ±  2%  +1.57% (p=0.002 n=6)
Query/sum_without_(l)(h_100),_instant_query/Mimir-8                                                                      9.232m ± 3%   9.452m ±  4%  +2.39% (p=0.041 n=6)
Query/sum_without_(l)(h_100),_instant_query/Prometheus-8                                                                 9.227m ± 1%   9.571m ±  2%  +3.72% (p=0.002 n=6)
Query/sum_without_(l)(h_100),_range_query_with_100_steps/Mimir-8                                                         17.14m ± 1%   17.51m ±  2%  +2.14% (p=0.004 n=6)
Query/sum_without_(l)(h_100),_range_query_with_100_steps/Prometheus-8                                                    17.61m ± 3%   17.47m ±  1%       ~ (p=0.065 n=6)
Query/sum_without_(l)(h_100),_range_query_with_1000_steps/Mimir-8                                                        87.56m ± 1%   87.51m ±  1%       ~ (p=0.818 n=6)
Query/sum_without_(l)(h_100),_range_query_with_1000_steps/Prometheus-8                                                   95.16m ± 8%   96.29m ±  7%       ~ (p=0.699 n=6)
Query/sum_without_(l)(h_2000),_instant_query/Mimir-8                                                                     166.9m ± 1%   168.5m ±  1%       ~ (p=0.093 n=6)
Query/sum_without_(l)(h_2000),_instant_query/Prometheus-8                                                                170.6m ± 1%   172.4m ±  2%  +1.09% (p=0.015 n=6)
Query/sum_without_(l)(h_2000),_range_query_with_100_steps/Mimir-8                                                        317.4m ± 2%   320.1m ±  1%       ~ (p=0.310 n=6)
Query/sum_without_(l)(h_2000),_range_query_with_100_steps/Prometheus-8                                                   354.3m ± 1%   355.1m ±  2%       ~ (p=0.240 n=6)
Query/sum_without_(l)(h_2000),_range_query_with_1000_steps/Mimir-8                                                        1.673 ± 1%    1.677 ±  1%       ~ (p=0.180 n=6)
Query/sum_without_(l)(h_2000),_range_query_with_1000_steps/Prometheus-8                                                   2.749 ± 2%    2.758 ±  2%       ~ (p=0.818 n=6)
Query/sum_without_(le)(h_1),_instant_query/Mimir-8                                                                       776.9µ ± 3%   801.4µ ±  2%  +3.16% (p=0.041 n=6)
Query/sum_without_(le)(h_1),_instant_query/Prometheus-8                                                                  782.3µ ± 1%   805.7µ ±  2%  +2.98% (p=0.004 n=6)
Query/sum_without_(le)(h_1),_range_query_with_100_steps/Mimir-8                                                          882.3µ ± 1%   895.2µ ±  3%       ~ (p=0.132 n=6)
Query/sum_without_(le)(h_1),_range_query_with_100_steps/Prometheus-8                                                     911.0µ ± 2%   905.3µ ±  2%       ~ (p=0.394 n=6)
Query/sum_without_(le)(h_1),_range_query_with_1000_steps/Mimir-8                                                         1.654m ± 1%   1.677m ±  1%  +1.39% (p=0.009 n=6)
Query/sum_without_(le)(h_1),_range_query_with_1000_steps/Prometheus-8                                                    1.721m ± 2%   1.728m ±  3%       ~ (p=0.485 n=6)
Query/sum_without_(le)(h_100),_instant_query/Mimir-8                                                                     9.273m ± 1%   9.499m ±  1%  +2.44% (p=0.002 n=6)
Query/sum_without_(le)(h_100),_instant_query/Prometheus-8                                                                9.404m ± 3%   9.652m ±  7%  +2.64% (p=0.041 n=6)
Query/sum_without_(le)(h_100),_range_query_with_100_steps/Mimir-8                                                        17.30m ± 4%   17.56m ±  1%       ~ (p=0.240 n=6)
Query/sum_without_(le)(h_100),_range_query_with_100_steps/Prometheus-8                                                   17.86m ± 1%   18.10m ±  7%  +1.35% (p=0.015 n=6)
Query/sum_without_(le)(h_100),_range_query_with_1000_steps/Mimir-8                                                       88.79m ± 1%   88.13m ±  1%       ~ (p=0.310 n=6)
Query/sum_without_(le)(h_100),_range_query_with_1000_steps/Prometheus-8                                                 101.19m ± 7%   96.79m ±  4%       ~ (p=0.310 n=6)
Query/sum_without_(le)(h_2000),_instant_query/Mimir-8                                                                    171.3m ± 1%   175.6m ±  2%  +2.48% (p=0.004 n=6)
Query/sum_without_(le)(h_2000),_instant_query/Prometheus-8                                                               173.0m ± 1%   174.5m ±  2%  +0.91% (p=0.009 n=6)
Query/sum_without_(le)(h_2000),_range_query_with_100_steps/Mimir-8                                                       325.6m ± 2%   330.3m ±  1%  +1.42% (p=0.041 n=6)
Query/sum_without_(le)(h_2000),_range_query_with_100_steps/Prometheus-8                                                  367.7m ± 1%   368.2m ±  1%       ~ (p=0.310 n=6)
Query/sum_without_(le)(h_2000),_range_query_with_1000_steps/Mimir-8                                                       1.719 ± 1%    1.716 ±  0%       ~ (p=0.699 n=6)
Query/sum_without_(le)(h_2000),_range_query_with_1000_steps/Prometheus-8                                                  2.931 ± 1%    2.935 ±  1%       ~ (p=0.818 n=6)
Query/sum(nh_1),_instant_query/Mimir-8                                                                                   807.7µ ± 2%   814.7µ ±  2%       ~ (p=0.093 n=6)
Query/sum(nh_1),_instant_query/Prometheus-8                                                                              810.6µ ± 1%   815.0µ ±  2%       ~ (p=0.065 n=6)
Query/sum(nh_1),_range_query_with_100_steps/Mimir-8                                                                      906.2µ ± 6%   912.4µ ±  2%       ~ (p=0.699 n=6)
Query/sum(nh_1),_range_query_with_100_steps/Prometheus-8                                                                 968.0µ ± 6%   987.9µ ±  3%       ~ (p=0.394 n=6)
Query/sum(nh_1),_range_query_with_1000_steps/Mimir-8                                                                     1.735m ± 5%   1.768m ±  3%       ~ (p=0.394 n=6)
Query/sum(nh_1),_range_query_with_1000_steps/Prometheus-8                                                                2.303m ± 5%   2.298m ±  5%       ~ (p=0.937 n=6)
Query/sum(nh_100),_instant_query/Mimir-8                                                                                 11.91m ± 1%   11.89m ±  0%       ~ (p=0.699 n=6)
Query/sum(nh_100),_instant_query/Prometheus-8                                                                            11.93m ± 1%   12.02m ±  4%       ~ (p=0.132 n=6)
Query/sum(nh_100),_range_query_with_100_steps/Mimir-8                                                                    21.69m ± 6%   21.97m ±  8%       ~ (p=0.132 n=6)
Query/sum(nh_100),_range_query_with_100_steps/Prometheus-8                                                               23.45m ± 6%   22.92m ±  4%       ~ (p=0.180 n=6)
Query/sum(nh_100),_range_query_with_1000_steps/Mimir-8                                                                   116.5m ± 6%   121.7m ±  7%       ~ (p=0.093 n=6)
Query/sum(nh_100),_range_query_with_1000_steps/Prometheus-8                                                              162.2m ± 4%   162.1m ±  2%       ~ (p=0.818 n=6)
Query/sum(nh_2000),_instant_query/Mimir-8                                                                                216.1m ± 1%   217.4m ±  1%       ~ (p=0.065 n=6)
Query/sum(nh_2000),_instant_query/Prometheus-8                                                                           216.7m ± 1%   218.5m ±  1%       ~ (p=0.093 n=6)
Query/sum(nh_2000),_range_query_with_100_steps/Mimir-8                                                                   422.1m ± 1%   432.1m ±  3%  +2.37% (p=0.009 n=6)
Query/sum(nh_2000),_range_query_with_100_steps/Prometheus-8                                                              539.6m ± 1%   544.5m ± 10%       ~ (p=0.310 n=6)
Query/sum(nh_2000),_range_query_with_1000_steps/Mimir-8                                                                   2.332 ± 2%    2.370 ±  3%       ~ (p=0.093 n=6)
Query/sum(nh_2000),_range_query_with_1000_steps/Prometheus-8                                                              4.087 ± 2%    4.079 ±  3%       ~ (p=1.000 n=6)
Query/sum_by_(l)(h_1),_instant_query/Mimir-8                                                                             792.5µ ± 2%   790.8µ ±  4%       ~ (p=1.000 n=6)
Query/sum_by_(l)(h_1),_instant_query/Prometheus-8                                                                        815.0µ ± 5%   802.2µ ±  1%       ~ (p=0.065 n=6)
Query/sum_by_(l)(h_1),_range_query_with_100_steps/Mimir-8                                                                893.3µ ± 3%   878.5µ ±  1%       ~ (p=0.093 n=6)
Query/sum_by_(l)(h_1),_range_query_with_100_steps/Prometheus-8                                                           920.6µ ± 1%   904.0µ ±  0%  -1.79% (p=0.002 n=6)
Query/sum_by_(l)(h_1),_range_query_with_1000_steps/Mimir-8                                                               1.651m ± 3%   1.664m ±  1%       ~ (p=0.310 n=6)
…